# static/specs/openapi/catalog/catalog_agent.yaml
openapi: 3.0.3
info:
  title: Rainbow Catalog Agent API - Entity Management
  description: OpenAPI specification for the Rainbow Catalog Agent Entity Management endpoints.
  version: 1.0.0
servers:
  - url: /v1/catalog-agent
    description: Base path for the catalog agent API
...
security:
  - BearerAuth: []
paths:
...
  /datasets:
    get:
      tags:
        - Datasets
      summary: Get all datasets
      operationId: getAllDatasets
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/PageParam"
      responses:
        "200":
          description: List of datasets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DatasetDto"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags:
        - Datasets
      summary: Create dataset
      operationId: createDataset
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewDatasetDto"
            example:
              catalogId: "urn:uuid:my-catalog-id"
              dctTitle: "My Dataset"
      responses:
        "200":
          description: Dataset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /datasets/catalog/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags:
        - Datasets
      summary: Get datasets by catalog ID
      operationId: getDatasetsByCatalogId
      responses:
        "200":
          description: List of datasets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DatasetDto"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /datasets/batch:
    post:
      tags:
        - Datasets
      summary: Get batch datasets
      operationId: getBatchDatasets
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchRequests"
      responses:
        "200":
          description: List of datasets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DatasetDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /datasets/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags:
        - Datasets
      summary: Get dataset by ID
      operationId: getDatasetById
      responses:
        "200":
          description: Dataset found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetDto"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      tags:
        - Datasets
      summary: Update dataset by ID
      operationId: updateDatasetById
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EditDatasetDto"
      responses:
        "200":
          description: Dataset updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetDto"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags:
        - Datasets
      summary: Delete dataset by ID
      operationId: deleteDatasetById
      responses:
        "202":
          description: Dataset deleted (Accepted)
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

  # --- Distributions ---
...
  /distributions/dataset/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags:
        - Distributions
      summary: Get distributions by dataset ID
      operationId: getDistributionsByDatasetId
      responses:
        "200":
          description: List of distributions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DistributionDto"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        format: int64
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        format: int64
...
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorInfo"
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorInfo"
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorInfo"
  schemas:
...
    ErrorInfo:
      type: object
      required:
        - message
        - error_code
        - cause
      properties:
        message:
          type: string
        error_code:
          type: integer
        details:
          type: string
        cause:
          type: string
      example:
        message: "Resource not found"
        error_code: 3200
        status_code: 404
        cause: "Catalog with ID urn:uuid:123 not found"

    # --- Common Wrappers ---
    BatchRequests:
      type: object
      properties:
        ids:
          type: array
          items:
            type: string
            format: urn
...
    DatasetDto:
      type: object
      properties:
        id:
          type: string
          format: urn
        dctConformsTo:
          type: string
        dctCreator:
          type: string
        dctIdentifier:
          type: string
        dctIssued:
          type: string
          format: date-time
        dctModified:
          type: string
          format: date-time
        dctTitle:
          type: string
        dctDescription:
          type: string
        catalogId:
          type: string
          format: urn
      example:
        {
          "catalogId": "urn:catalog:4958f37d-17cf-44e7-ae18-1f26b6fe15ff",
          "dctConformsTo": null,
          "dctCreator": null,
          "dctDescription": null,
          "dctIdentifier": "urn:dataset:10825ca3-a75e-43bd-9a82-603759cb8dec",
          "dctIssued": "2026-01-10T14:23:29.091150Z",
          "dctModified": null,
          "dctTitle": "superfoaf",
          "id": "urn:dataset:10825ca3-a75e-43bd-9a82-603759cb8dec"
        }
    NewDatasetDto:
      type: object
      required:
        - catalogId
      properties:
        id:
          type: string
          format: urn
        dctConformsTo:
          type: string
        dctCreator:
          type: string
        dctTitle:
          type: string
        dctDescription:
          type: string
        catalogId:
          type: string
          format: urn
      example:
        {
          "catalogId": "urn:catalog:4958f37d-17cf-44e7-ae18-1f26b6fe15ff",
          "dctConformsTo": null,
          "dctCreator": null,
          "dctDescription": null,
          "dctTitle": "superfoaf",
        }
    EditDatasetDto:
      type: object
      properties:
        dctConformsTo:
          type: string
        dctCreator:
          type: string
        dctTitle:
          type: string
        dctDescription:
          type: string
      example:
        {
          "dctConformsTo": null,
          "dctCreator": null,
          "dctDescription": null,
          "dctTitle": "superfoaf",
        }

    # --- Distributions ---
    DistributionDto:
      type: object
      properties:
        id:
          type: string
          format: urn
        dctIssued:
          type: string
          format: date-time
        dctModified:
          type: string
          format: date-time
        dctTitle:
          type: string
        dctDescription:
          type: string
        dctFormat:
          type: string
          description: "Format in protocol+action format"
        dcatAccessService:
          type: string
        datasetId:
          type: string
          format: urn
      example:
        {
          "datasetId": "urn:dataset:10825ca3-a75e-43bd-9a82-603759cb8dec",
          "dcatAccessService": "urn:data-service:37840ca3-79f6-4542-8394-0c57dd93650f",
          "dctDescription": null,
          "dctFormat": "Http+Push",
          "dctIssued": "2026-01-10T14:24:29.424022Z",
          "dctModified": null,
          "dctTitle": "superfoaf",
          "id": "urn:distribution:77116828-c4dd-4632-8946-969bb7903819"
        }
...